+++ starseqr_utils/__init__.py
@@ -7,9 +7,10 @@ import starseqr_utils.support_funcs_rna
 import starseqr_utils.run_primer3
 import starseqr_utils.star_funcs
 import starseqr_utils.cross_homology
+import starseqr_utils.homology_graph
 import starseqr_utils.overhang_diversity
 import starseqr_utils.run_assembly
 import starseqr_utils.annotate_db
 import starseqr_utils.sv2bedpe
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"